feat: extract HUD settings panel
